Feat/before pull (#7)
