+++ src/components/comments/Comments.tsx
@@ -3,76 +3,27 @@ import * as S from './Styled.Comments';
 import { fetchComments } from '../../services/supabase/comments';
 import { useQuery } from '@tanstack/react-query';
 import { useParams } from 'react-router';
-import { Comment } from '../../types/supabase';
-import useCommentMutation from '../../hooks/useCommentMutation';
+import ReplyComments from './ReplyComments';
 import CommentForm from './CommentForm';
-import CommentsPanel from './CommentPanel';
+import CommentItem from './CommentItem';
 
+import type { Comment } from '../../types/supabase';
 const Comments = () => {
   // login 완료되면 수정하기
-  const uid = '3';
+  const uid = '23';
   const { id: pid } = useParams() as { id: string };
 
   // 댓글 작성 버튼 컨트롤
   const [isCommenting, setIsCommenting] = useState(false);
 
-  // 댓글 수정 state들
-  const [updateCommentId, setUpdateCommentId] = useState<number | null>(0);
-  const [updateComment, setUpdateComment] = useState('');
+  // 댓글 수정
   const [isUpdating, setIsUpdating] = useState(false);
 
-  // useMutation Hooks
-  const { updateCommentMutation, deleteCommentMutation } = useCommentMutation();
+  // 대댓글 보기
+  const [isViewingReply, setIsViewingReply] = useState(false);
 
   const handleCommentFormBtnClick = () => {
     setIsCommenting(!isCommenting);
-  };
-
-  const handleUpdateCommentBtnClick = (cid: number, comment: string) => {
-    setIsUpdating(true);
-    setUpdateComment(comment);
-    setUpdateCommentId(cid);
-  };
-
-  const handleUpdateCommentCancel = () => {
-    setIsUpdating(false);
-    setUpdateCommentId(null);
-  };
-
-  const handleUpdateCommentInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setUpdateComment(e.target.value);
-  };
-
-  const handleUpdateClickBtn = () => {
-    if (updateComment === '') {
-      alert('작성된 댓글이 없습니다.');
-      return false;
-    }
-
-    const newComment = {
-      cid: updateCommentId!,
-      body: updateComment
-    };
-
-    updateCommentMutation.mutate(newComment);
-
-    setUpdateComment('');
-    setIsUpdating(false);
-    setUpdateCommentId(null);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      handleUpdateClickBtn();
-    }
-  };
-
-  const handleDeleteCommentBtnClick = (cid: number) => {
-    const isConfirmed = window.confirm('삭제하시겠습니까?');
-    if (!isConfirmed) {
-      return false;
-    }
-    deleteCommentMutation.mutate(cid);
   };
 
   const defaultQueryOptions = {
@@ -105,35 +56,10 @@ const Comments = () => {
       <S.CommentsHr />
       <S.CommentList>
         {comments?.map((comment) => (
-          <S.CommentItem key={comment.cid}>
-            <S.CommentProfileImgBox>사진</S.CommentProfileImgBox>
-            <S.CommentTextBox>
-              <S.CommentAuthor>{comment.nickname}</S.CommentAuthor>
-              {isUpdating && updateCommentId == comment.cid ? (
-                <S.CommentInput
-                  type="text"
-                  value={updateComment}
-                  onChange={handleUpdateCommentInputChange}
-                  onKeyDown={handleKeyDown}
-                />
-              ) : (
-                <S.CommentBody>{comment.body}</S.CommentBody>
-              )}
-            </S.CommentTextBox>
-            {uid === comment.uid && isUpdating && updateCommentId == comment.cid ? (
-              <CommentsPanel
-                commenting={true}
-                handleUpdateClickBtn={handleUpdateClickBtn}
-                handleUpdateCommentCancel={handleUpdateCommentCancel}
-              />
-            ) : (
-              <CommentsPanel
-                commenting={false}
-                handleUpdateCommentBtnClick={() => handleUpdateCommentBtnClick(comment.cid, comment.body)}
-                handleDeleteCommentBtnClick={() => handleDeleteCommentBtnClick(comment.cid)}
-              />
-            )}
-          </S.CommentItem>
+          <React.Fragment key={comment.cid}>
+            <CommentItem comment={comment} uid={uid} isUpdating={isUpdating} setIsUpdating={setIsUpdating} />
+            <ReplyComments isViewingReply={isViewingReply} setIsViewingReply={setIsViewingReply} />
+          </React.Fragment>
         ))}
       </S.CommentList>
       {isCommenting && <CommentForm uid={uid} pid={pid} />}
