+++ src/components/comments/ReplyComments.tsx
@@ -1,7 +1,22 @@
-import React from 'react';
+import { SetStateAction } from 'jotai';
+import React, { Dispatch } from 'react';
 
-const ReplyComments = () => {
-  return <div>ReplyComments</div>;
+interface ReplyCommentsProps {
+  isViewingReply: boolean;
+  setIsViewingReply: Dispatch<SetStateAction<boolean>>;
+}
+
+const ReplyComments = ({ isViewingReply, setIsViewingReply }: ReplyCommentsProps) => {
+  return isViewingReply ? (
+    <>
+      <div onClick={() => setIsViewingReply(!isViewingReply)}>닫기</div>
+      <span>댓글입니다</span>
+    </>
+  ) : (
+    <>
+      <div onClick={() => setIsViewingReply(!isViewingReply)}>열기</div>
+    </>
+  );
 };
 
 export default ReplyComments;
